Refactor account switching logic into custom hook
Moved account switching logic from App.tsx to a new useAccountSwitching hook for better separation of concerns and reusability. This change simplifies the main app component and encapsulates related state and handlers.
