+++ src/App.tsx
@@ -95,7 +95,7 @@ import { useGitCommitController } from "./features/app/hooks/useGitCommitControl
 import { WorkspaceHome } from "./features/workspaces/components/WorkspaceHome";
 import { useWorkspaceHome } from "./features/workspaces/hooks/useWorkspaceHome";
 import { useWorkspaceAgentMd } from "./features/workspaces/hooks/useWorkspaceAgentMd";
-import { cancelCodexLogin, pickWorkspacePath, runCodexLogin } from "./services/tauri";
+import { pickWorkspacePath } from "./services/tauri";
 import type {
   AccessMode,
   ComposerEditorSettings,
@@ -104,6 +104,7 @@ import type {
 import { OPEN_APP_STORAGE_KEY } from "./features/app/constants";
 import { useOpenAppIcons } from "./features/app/hooks/useOpenAppIcons";
 import { useCodeCssVars } from "./features/app/hooks/useCodeCssVars";
+import { useAccountSwitching } from "./features/app/hooks/useAccountSwitching";
 
 const AboutView = lazy(() =>
   import("./features/about/components/AboutView").then((module) => ({
@@ -631,11 +632,18 @@ function MainApp() {
     customPrompts: prompts,
     onMessageActivity: queueGitStatusRefresh
   });
-  const [accountSwitching, setAccountSwitching] = useState(false);
-  const accountSwitchCanceledRef = useRef(false);
-  const activeAccount = activeWorkspaceId
-    ? accountByWorkspace[activeWorkspaceId] ?? null
-    : null;
+  const {
+    activeAccount,
+    accountSwitching,
+    handleSwitchAccount,
+    handleCancelSwitchAccount,
+  } = useAccountSwitching({
+    activeWorkspaceId,
+    accountByWorkspace,
+    refreshAccountInfo,
+    refreshAccountRateLimits,
+    alertError,
+  });
   const activeThreadIdRef = useRef<string | null>(activeThreadId ?? null);
   const { getThreadRows } = useThreadRows(threadParentById);
   useEffect(() => {
@@ -1115,62 +1123,6 @@ function MainApp() {
     },
     [activeWorkspace, connectWorkspace, sendUserMessageToThread, startThreadForWorkspace],
   );
-
-  const isCodexLoginCanceled = useCallback((error: unknown) => {
-    const message =
-      typeof error === "string" ? error : error instanceof Error ? error.message : "";
-    const normalized = message.toLowerCase();
-    return (
-      normalized.includes("codex login canceled") ||
-      normalized.includes("codex login cancelled") ||
-      normalized.includes("request canceled")
-    );
-  }, []);
-
-  const handleSwitchAccount = useCallback(async () => {
-    if (!activeWorkspaceId || accountSwitching) {
-      return;
-    }
-    accountSwitchCanceledRef.current = false;
-    setAccountSwitching(true);
-    try {
-      await runCodexLogin(activeWorkspaceId);
-      if (accountSwitchCanceledRef.current) {
-        return;
-      }
-      await refreshAccountInfo(activeWorkspaceId);
-      await refreshAccountRateLimits(activeWorkspaceId);
-    } catch (error) {
-      if (accountSwitchCanceledRef.current || isCodexLoginCanceled(error)) {
-        return;
-      }
-      alertError(error);
-    } finally {
-      setAccountSwitching(false);
-      accountSwitchCanceledRef.current = false;
-    }
-  }, [
-    activeWorkspaceId,
-    accountSwitching,
-    refreshAccountInfo,
-    refreshAccountRateLimits,
-    alertError,
-    isCodexLoginCanceled,
-  ]);
-
-  const handleCancelSwitchAccount = useCallback(async () => {
-    if (!activeWorkspaceId || !accountSwitching) {
-      return;
-    }
-    accountSwitchCanceledRef.current = true;
-    try {
-      await cancelCodexLogin(activeWorkspaceId);
-    } catch (error) {
-      alertError(error);
-    } finally {
-      setAccountSwitching(false);
-    }
-  }, [activeWorkspaceId, accountSwitching, alertError]);
 
 
   const handleCreatePrompt = useCallback(
